1.impl concurrency for paralle compute,
2.made sure memory cleanup and compute cancellation happens on cancellation from the frontend

3. using uint8 for r,g,b and using buffer bufIO to write voxels into .obj for more perf

4. added more doc, and cleanup here and there

+++ backend/types.go
@@ -5,7 +5,7 @@ import "image"
 type Vector3 struct{ x, y, z int }
 
 type RGB struct {
-	r, g, b float64
+	r, g, b uint8
 }
 
 type Mode int
